requirements.txt
pandas
numpy
matplotlib
Pillow
tensorflow
streamlit
plotly
reportlab
requests